ch2/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
dotenv==0.9.9
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.14.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
websockets==15.0.1
